Moved many generic methods from TeamCommandsUtils to a new CommandUtils
